Move skeleton namespace
