+++ src/main/java/dev/slimevr/posestreamer/PoseDataStream.java
@@ -6,7 +6,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 
-import io.eiren.vr.processor.HumanSkeleton;
+import dev.slimevr.vr.processor.skeleton.HumanSkeleton;
 
 public abstract class PoseDataStream implements AutoCloseable {
