+++ src/main/java/dev/slimevr/vr/processor/skeleton/SkeletonConfig.java
@@ -1,4 +1,4 @@
-package dev.slimevr.vr.processor;
+package dev.slimevr.vr.processor.skeleton;
 
 import java.util.EnumMap;
 import java.util.Map;
