+++ src/main/java/dev/slimevr/vr/processor/skeleton/SkeletonNodeOffset.java
@@ -1,4 +1,4 @@
-package dev.slimevr.vr.processor;
+package dev.slimevr.vr.processor.skeleton;
 
 public enum SkeletonNodeOffset {
