+++ src/main/java/io/eiren/vr/VRServer.java
@@ -20,13 +20,13 @@ import dev.slimevr.bridge.NamedPipeBridge;
 import dev.slimevr.bridge.SteamVRPipeInputBridge;
 import dev.slimevr.bridge.VMCBridge;
 import dev.slimevr.bridge.WebSocketVRBridge;
+import dev.slimevr.vr.processor.skeleton.HumanSkeleton;
 import io.eiren.util.OperatingSystem;
 import io.eiren.util.ann.ThreadSafe;
 import io.eiren.util.ann.ThreadSecure;
 import io.eiren.util.ann.VRServerThread;
 import io.eiren.util.collections.FastList;
 import io.eiren.vr.processor.HumanPoseProcessor;
-import io.eiren.vr.processor.HumanSkeleton;
 import io.eiren.vr.trackers.HMDTracker;
 import io.eiren.vr.trackers.ShareableTracker;
 import io.eiren.vr.trackers.TrackersUDPServer;
